Modifiche fondamentali per proseguire

+++ src/main/java/it/generationitaly/dao/AnnuncioDAO.java
@@ -9,6 +9,6 @@ public interface AnnuncioDAO {
 
 	List<Annuncio> findAll(Connection connection) throws DAOException;
 	
-	List<Annuncio> findFiltered(Connection connection, String marca, String modello, double prezzo) throws DAOException;
+	List<Annuncio> findFiltered(Connection connection, String marca, String modello, int prezzo) throws DAOException;
 
 }
