+++ src/main/java/it/generationitaly/service/AnnuncioService.java
@@ -8,5 +8,5 @@ public interface AnnuncioService {
 	
 	List<Annuncio> findAll() throws ServiceException;
 	
-	List<Annuncio> findFiltered(String marca, String modello, double prezzo) throws ServiceException;
+	List<Annuncio> findFiltered(String marca, String modello, int prezzo) throws ServiceException;
 }
